+++ src/components/reusable-ui/Button.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { theme } from "../../../theme";
+import { theme } from "../../theme";
 
 function Button({ label, Icon }) {
   return (
